+++ src/platform/ui_ncurses.c
@@ -208,7 +208,7 @@ char *ui_platform_show_open_dialog(Window *parent, const FileDialogParams *param
     return NULL;
 }
 
-char *ui_platform_show_save_dialog(Window *parent, const FileDialogParams *params) {
+char *ui_platform_show_save_dialog(Window *parent, FileDialogParams *params) {
     (void) parent;
     (void) params;
     return NULL;